Updating a document now does a diff between the previous indexes and the new indexes to make sure only changed indexes are updated.
This is significant as we expect many document updates to change none or
only some of the indexes. The diff isn't particularly efficient, but it
happens outside of any locks and should be happening over a relatively
small array. If it did prove a problem, switching the meta's indexes to a
set would be simple and effective.