Ranged queries can not be done against the sort index, provided the sort index is a dynamic sort index.
The fact that we have 4 different index implementation, and 3 different query
paths is a pain. If everything was a "dynamic sort index" (aka, some type of
tree), things would get simpler and more powerful. Nevertheless, there's
some serious memory-win by creating a static index when you know you won't
be doing ranged base queries against it.

What would be nice is being able to know the # of elements within a range.
I believe this is doable with an indexable skiplist, but whether there's any
real value in this needs to be tested. (the benefit would be to direct
more calls to findBySort by the range is small, but since calls which go to
findByIndex are already small, I'm not sure what real benefit this would have).