This branch is focused on two things: simplifying the indexes and supporting richer queries. To this end, only the skiplist index remains, and there's no longer a distinction between sorting indexes and filtering indexes.
As a consequence, the following things have been removed:

- Query cache
- Efficient filtering
- Append/Prepend
- Bulk index load
- Aggregation

Efficient filtering is going to come back in the next commit. Query cache
should come back shortly thereafter.

The other three are doable, but I'll wait until they become necessary.

Documentation is also out of sync, and a lot of tests are broken.

I still think there's value in having a lightweight specialized index which
represents a pre-filtered state. But, for now anyways, we'll skip that.